+++ docs/sphinx/changelog.rst
@@ -14,6 +14,8 @@ New Features
 - Add an option to construct the stiffness tensor from the full 81 components (:issue:`11`, :merge:`23`). By `Kyle
   Brindley`_.
 - Add an option to construct and rotate the stiffness tensor (:issue:`12`, :merge:`24`). By `Kyle Brindley`_.
+- Add an energy and derivatives overload that accepts an Euler angle rotation for the stiffness matrix (:issue:`14`,
+  :merge:`27`). By `Kyle Brindley`_.
 
 Internal Changes
 ================